+++ test/lightning/projects/provisioner_test.exs
@@ -435,7 +435,7 @@ defmodule Lightning.Projects.ProvisionerTest do
                workflows: [
                  %{
                    jobs: [
-                     %{id: ["This email address already exists."]},
+                     %{id: ["This value should be unique."]},
                      %{},
                      %{}
                    ]
@@ -458,7 +458,7 @@ defmodule Lightning.Projects.ProvisionerTest do
       assert flatten_errors(changeset) == %{
                workflows: [
                  %{
-                   triggers: [%{id: ["This email address already exists."]}, %{}]
+                   triggers: [%{id: ["This value should be unique."]}, %{}]
                  }
                ]
              }
@@ -481,7 +481,7 @@ defmodule Lightning.Projects.ProvisionerTest do
                workflows: [
                  %{
                    edges: [
-                     %{id: ["This email address already exists."]},
+                     %{id: ["This value should be unique."]},
                      %{},
                      %{}
                    ]