+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to
 
 ### Added
 
+- Add workflows API to create, update, get and list.
+  [#1887](https://github.com/OpenFn/lightning/issues/1887)
 - Show email address of credential owner on project credentials page
   [#2210](https://github.com/OpenFn/lightning/issues/2210)